work on indexer and iteration
