+++ crates/middleware/src/command/mod.rs
@@ -1,0 +1,1 @@
+pub mod ripgrep;
