+++ crates/middleware/src/indexer/mod.rs
@@ -1,10 +1,7 @@
 use ahash::AHashMap;
-use serde::Deserialize;
-use serde_json as json;
 use std::collections::hash_map::DefaultHasher;
 use std::hash::{Hash, Hasher};
 use std::path::Path;
-use std::time;
 use terraphim_config::{ConfigState, ServiceType};
 use terraphim_types::{Index, SearchQuery};
 
@@ -15,138 +12,6 @@ mod ripgrep;
 
 pub use logseq::LogseqIndexer;
 pub use ripgrep::RipgrepIndexer;
-
-#[derive(Clone, Debug, Deserialize, PartialEq, Eq, Hash)]
-#[serde(tag = "type", content = "data")]
-#[serde(rename_all = "snake_case")]
-pub enum Message {
-    Begin(Begin),
-    End(End),
-    Match(Match),
-    Context(Context),
-    Summary(Summary),
-}
-
-/// The `Begin` message is sent at the beginning of each search.
-/// It contains the path that is being searched, if one exists.
-#[derive(Clone, Debug, Deserialize, PartialEq, Eq, Hash)]
-pub struct Begin {
-    pub path: Option<Data>,
-}
-
-impl Begin {
-    /// Gets the path of the file being searched (if set).
-    pub(crate) fn path(&self) -> Option<String> {
-        as_path(&self.path)
-    }
-}
-
-/// The `End` message is sent at the end of a search.
-#[derive(Clone, Debug, Deserialize, PartialEq, Eq, Hash)]
-pub struct End {
-    path: Option<Data>,
-    binary_offset: Option<u64>,
-    stats: Stats,
-}
-
-/// The `Summary` message is sent at the end of a search.
-#[derive(Clone, Debug, Deserialize, PartialEq, Eq, Hash)]
-pub struct Summary {
-    elapsed_total: Duration,
-    stats: Stats,
-}
-
-/// The `Match` message is sent for each non-overlapping match of a search.
-#[derive(Clone, Debug, Deserialize, PartialEq, Eq, Hash)]
-pub struct Match {
-    pub path: Option<Data>,
-    pub lines: Data,
-    line_number: Option<u64>,
-    absolute_offset: u64,
-    pub submatches: Vec<SubMatch>,
-}
-
-impl Match {
-    /// Gets the path of the file being searched (if set).
-    pub(crate) fn path(&self) -> Option<String> {
-        as_path(&self.path)
-    }
-}
-
-/// The `Context` specifies the lines surrounding a match.
-#[derive(Clone, Debug, Deserialize, PartialEq, Eq, Hash)]
-pub struct Context {
-    pub path: Option<Data>,
-    pub lines: Data,
-    line_number: Option<u64>,
-    absolute_offset: u64,
-    submatches: Vec<SubMatch>,
-}
-
-impl Context {
-    /// Gets the path of the file being searched (if set).
-    pub(crate) fn path(&self) -> Option<String> {
-        as_path(&self.path)
-    }
-}
-
-/// A `SubMatch` is a match within a match.
-#[derive(Clone, Debug, Deserialize, PartialEq, Eq, Hash)]
-pub struct SubMatch {
-    #[serde(rename = "match")]
-    m: Data,
-    start: usize,
-    end: usize,
-}
-
-/// The `Data` type is used for fields that may be either text or bytes.
-/// In contains the raw data of the field.
-#[derive(Clone, Debug, Deserialize, PartialEq, Eq, Hash)]
-#[serde(untagged)]
-pub enum Data {
-    Text { text: String },
-    // This variant is used when the data isn't valid UTF-8. The bytes are
-    // base64 encoded, so using a String here is OK.
-    Bytes { bytes: String },
-}
-
-/// Gets the path from a `Data` type.
-fn as_path(data: &Option<Data>) -> Option<String> {
-    // Return immediately if the data is None
-    let data = match data {
-        Some(data) => data,
-        None => return None,
-    };
-    match data {
-        Data::Text { text } => Some(text.clone()),
-        _ => None,
-    }
-}
-
-#[derive(Clone, Debug, Deserialize, PartialEq, Eq, Hash)]
-struct Stats {
-    elapsed: Duration,
-    searches: u64,
-    searches_with_match: u64,
-    bytes_searched: u64,
-    bytes_printed: u64,
-    matched_lines: u64,
-    matches: u64,
-}
-
-#[derive(Clone, Debug, Deserialize, PartialEq, Eq, Hash)]
-struct Duration {
-    #[serde(flatten)]
-    duration: time::Duration,
-    human: String,
-}
-
-/// Decode JSON Lines into a Vec<Message>.
-pub fn json_decode(jsonlines: &str) -> Result<Vec<Message>> {
-    Ok(json::Deserializer::from_str(jsonlines)
-        .into_iter()
-        .collect::<std::result::Result<Vec<Message>, serde_json::Error>>()?)
-}
 
 fn calculate_hash<T: Hash>(t: &T) -> String {
     let mut s = DefaultHasher::new();
