+++ crates/middleware/src/lib.rs
@@ -1,6 +1,7 @@
 use serde_json as json;
 use terraphim_config::TerraphimConfigError;
 
+mod command;
 pub mod graph_builder;
 pub mod indexer;
